Remove place module and  Integrate map modules
