+++ src/redux/modules/map.js
@@ -1,6 +1,11 @@
-import {call, put, takeLatest, all} from 'redux-saga/effects';
-import {ViewPort, Message} from 'redux/models';
+import {call, put, take, cancelled, cancel, takeLatest, fork, all} from 'redux-saga/effects';
+import {delay} from 'redux-saga'
+import {Coordinate, ViewPort, Message, MapStatus} from 'redux/models';
+import {List} from 'immutable';
 import geoLocation from 'helpers/geoLocation';
+import API from 'helpers/api';
+import * as storage from 'helpers/storage';
+import {isMove, setNextTime} from 'helpers/distance';
 
 const GET_CURRENT_LOCATION = 'barry/map/GET_CURRENT_LOCATION';
 const GET_CURRENT_LOCATION_SUCCESS = 'barry/map/GET_CURRENT_LOCATION_SUCCESS';
@@ -8,8 +13,18 @@ const GET_CURRENT_LOCATION_FAIL = 'barry/map/GET_CURRENT_LOCATION_FAIL';
 const SET_VIEWPORT = 'barry/map/SET_VIEWPORT';
 const SET_VIEWPORT_SUCCESS = 'barry/map/SET_VIEWPORT_SUCCESS';
 const SET_VIEWPORT_FAIL = 'barry/map/SET_VIEWPORT_FAIL';
+const GET_COORDINATES = 'barry/map/GET_COORDINATES';
+const GET_COORDINATES_SUCCESS = 'barry/map/GET_COORDINATES_SUCCESS';
+const GET_COORDINATES_FAIL = 'barry/map/GET_COORDINATES_FAIL';
+const POST_COORDINATE = 'barry/map/POST_COORDINATE';
+const POST_COORDINATE_SUCCESS = 'barry/map/POST_COORDINATE_SUCCESS';
+const POST_COORDINATE_FAIL = 'barry/map/POST_COORDINATE_FAIL';
+const START_UPDATE_POSITION = 'barry/map/START_UPDATE_POSITION';
+const STOP_UPDATE_POSITION = 'barry/map/END_UPDATE_POSITION';
 
 const initialState = {
+  coordinates: new List(),
+  mapStatus: new MapStatus(),
   viewport: new ViewPort(),
   message: new Message()
 };
@@ -20,8 +35,8 @@ export default function reducer(state = initialState, action = {}) {
       return state;
     case GET_CURRENT_LOCATION_SUCCESS:
       return {
-        viewport: new ViewPort(action.viewport),
-        message: state.message
+        ...state,
+        viewport: new ViewPort(action.viewport)
       };
     case GET_CURRENT_LOCATION_FAIL:
       return state;
@@ -29,10 +44,51 @@ export default function reducer(state = initialState, action = {}) {
       return state;
     case SET_VIEWPORT_SUCCESS:
       return {
+        ...state,
         viewport: new ViewPort(action.viewport)
       };
     case SET_VIEWPORT_FAIL:
       return state;
+    case GET_COORDINATES:
+      return state;
+    case GET_COORDINATES_SUCCESS:
+      const coordinates = List(action.coordinates.map((v) => {
+        return new Coordinate(v);
+      }));
+      console.log(coordinates);
+      return {
+        ...state,
+        coordinates: coordinates,
+      };
+    case GET_COORDINATES_FAIL:
+      return {
+        ...state,
+        message: state.message.set("error", action.error)
+      };
+    case POST_COORDINATE:
+      return state;
+    case POST_COORDINATE_SUCCESS:
+      return {
+        ...state,
+        coordinates: state.coordinates.push(
+          new Coordinate(action.coordinate)
+        ),
+      };
+    case POST_COORDINATE_FAIL:
+      return {
+        ...state,
+        message: state.message.set("error", action.error)
+      };
+    case START_UPDATE_POSITION:
+      return {
+        ...state,
+        mapStatus: state.mapStatus.set("updatingPostion", true)
+      };
+    case STOP_UPDATE_POSITION:
+      return {
+        ...state,
+        mapStatus: state.mapStatus.set("updatingPostion", false)
+      };
     default:
       return state;
   }
@@ -48,6 +104,32 @@ export function setViewPort(viewport) {
 export function getCurrentLocation() {
   return {
     type: GET_CURRENT_LOCATION
+  };
+}
+
+export function getCoordinates(payload) {
+  return {
+    type: GET_COORDINATES,
+    payload: payload
+  };
+}
+
+export function postCoordinate(payload) {
+  return {
+    type: POST_COORDINATE,
+    payload: payload
+  }
+}
+
+export function startUpdatePosition() {
+  return {
+    type: START_UPDATE_POSITION
+  };
+}
+
+export function stopUpdatePosition() {
+  return {
+    type: STOP_UPDATE_POSITION
   };
 }
 
@@ -82,5 +164,6 @@ export function *mapSagas() {
   yield all([
     takeLatest(SET_VIEWPORT, hundleSetViewPort),
     takeLatest(GET_CURRENT_LOCATION, hundleGetCurrentLocation),
+    takeLatest(GET_COORDINATES, hundleGetCoordinates)
   ]);
 }
