+++ src/redux/reducers.js
@@ -1,9 +1,7 @@
 import auth from './modules/auth';
 import map from './modules/map';
-import place from './modules/place';
 
 export const reducers = {
   auth,
-  map,
-  place
+  map
 };
